Changed arguments of functions to const when appropriate, made some changes to avoid Warinings(Made sure the project runs successfully against all given input files)

+++ Ex1/Father/create_process.c
@@ -36,7 +36,7 @@ long create_process_simple_main(char *son_str) {
 	GetExitCodeProcess(proc_info.hProcess, &exitcode);
 	CloseHandle(proc_info.hProcess); /* Closing the handle to the process */
 	CloseHandle(proc_info.hThread); /* Closing the handle to the main thread of the process */
-	return (long)exitcode;
+	return (int)exitcode;
 }
 
 BOOL create_process_simple(LPTSTR command_line, PROCESS_INFORMATION *p_info, TCHAR *son_str) {
